refactor(schemas): move the storable-secret-type boundary out of the schema module (#14974)
The narrowing annotation and its validator pushed schemas_system.py to 4366
lines, 40 over a ceiling a grandfathered file may not cross. They move to
api/schemas_secrets.py -- a self-contained pair with no other tie to the several
hundred models around them -- and schemas_system imports StorableSecretType from
there, so api.schemas_system.StorableSecretType resolves for both field
declarations exactly as before. No re-export shim: the plain import is the
re-export. 4366 -> 4309, ceiling lowered to 4309 in both copies.

Nothing was deleted to reach the number. The wildcard is still rejected at
runtime and the advertised schema still narrows to SecretType.concrete(),
verified against both models.
